+++ demo/src/components/controls/radio.tsx
@@ -26,7 +26,7 @@ const items = [
 } as WUP.Radio.Options["validations"];
 
 (window as any).storedRadioItems = {
-  items: items.slice(0, 4),
+  items,
 };
 
 export default function RadioControlView() {
@@ -64,27 +64,29 @@ export default function RadioControlView() {
         <wup-radio
           ref={(el) => {
             if (el) {
+              el.$options.name = "test"; // todo remove the control
+              setTimeout(() => {
+                const _items = [
+                  { text: "a1", value: { name: "Dik" } },
+                  { text: "a2", value: { name: "Yomma" } },
+                ];
+                el.$options.items = _items;
+                setTimeout(() => {
+                  el.$value = _items[1].value;
+                  console.warn(_items[0].value === el.$options.items[0].value);
+                  // el.$value = _items[0].value;
+                }, 1);
+              }, 300);
+            }
+          }}
+        />
+        <wup-radio
+          ref={(el) => {
+            if (el) {
               el.$options.name = "disabled";
               el.$options.disabled = true;
-              el.$options.items = items;
-            }
-          }}
-        />
-        <wup-radio
-          ref={(el) => {
-            if (el) {
-              el.$options.name = "readonly";
-              el.$options.items = items;
-              el.$options.readOnly = true;
-            }
-          }}
-        />
-        <wup-radio
-          ref={(el) => {
-            if (el) {
-              el.$options.name = "withInitValue";
-              el.$options.items = items;
-              el.$initValue = ir - 2;
+              el.$options.items = items.slice(0, 4);
+              el.$initValue = el.$options.items[1].value;
             }
           }}
         />
@@ -92,8 +94,16 @@ export default function RadioControlView() {
           initValue="13"
           ref={(el) => {
             if (el) {
+              el.$options.name = "readonly";
+              el.$options.items = items.slice(0, 4);
+            }
+          }}
+        />
+        <wup-radio
+          ref={(el) => {
+            if (el) {
               el.$options.name = "reversed";
-              el.$options.items = items;
+              el.$options.items = items.slice(0, 4);
               el.$options.reverse = true;
             }
           }}